requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
alembic==1.12.1
psycopg2-binary>=2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
kafka-python>=2.0.2
pydantic>=2.10.0
pydantic-settings>=2.6.0
slowapi==0.1.9

# Testing dependencies
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.0
httpx==0.27.0
faker==28.4.1

